+++ build.gradle
@@ -30,7 +30,7 @@ ext {
 
 dependencies {
 	//restdocs-dsl
-	implementation 'io.github.lhs8701:restdocs-dsl:1.2.1'
+	implementation 'io.github.lhs8701:restdocs-dslized-library:1.2.1'
 
 	//test lombok
 	testCompileOnly 'org.projectlombok:lombok'
