--- src/test/java/orgo/backend/TestForTest.java
@@ -1,79 +1,0 @@
-package orgo.backend;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import io.jsonwebtoken.Claims;
-import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.SignatureAlgorithm;
-import io.jsonwebtoken.security.Keys;
-import jakarta.transaction.Transactional;
-import lombok.extern.slf4j.Slf4j;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.boot.test.autoconfigure.restdocs.AutoConfigureRestDocs;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.http.MediaType;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.test.context.ActiveProfiles;
-import org.springframework.test.web.servlet.MockMvc;
-import org.springframework.test.web.servlet.ResultActions;
-import orgo.backend.domain._2user.dao.UserRepository;
-import orgo.backend.domain._2user.domain.User;
-import orgo.backend.domain._etc.test.TestController;
-import orgo.backend.global.config.security.JwtAuthenticationToken;
-
-import javax.crypto.SecretKey;
-import java.nio.charset.StandardCharsets;
-import java.security.Key;
-import java.util.Date;
-import java.util.UUID;
-import java.util.stream.Collectors;
-
-import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.document;
-import static org.springframework.restdocs.payload.PayloadDocumentation.*;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-import static hansol.restdocsdsl.element.HeaderElement.header;
-
-
-@SpringBootTest
-@ActiveProfiles("inttest")
-@AutoConfigureMockMvc
-@AutoConfigureRestDocs
-@Transactional
-@Slf4j
-public class TestForTest {
-    @Autowired
-    protected MockMvc mvc;
-    @Autowired
-    protected ObjectMapper objectMapper;
-
-    @Autowired
-    UserRepository userRepository;
-
-    @Test
-    @DisplayName("")
-    void test() throws Exception {
-        // given
-        TestController.TestRequest requestDto = new TestController.TestRequest("test");
-        //when
-        ResultActions actions = mvc.perform(post("/test")
-                .contentType(MediaType.APPLICATION_JSON)
-                .content(objectMapper.writeValueAsString(requestDto)));
-
-        //then
-        actions.andExpect(status().isOk())
-                .andDo(docs("test",
-                        requestFields(
-                                fieldWithPath("field").description("테스트 요청 필드")
-                                ),
-                        responseFields(
-                                fieldWithPath("field").description("테스트 응답 필드")
-                        )));
-    }
-}
